Move scripts into the module
