+++ jellyfin_migrator/__main__.py
@@ -1,0 +1,10 @@
+#!/usr/bin/env python3
+
+
+if __name__ == '__main__':
+    """
+    CommandLine:
+        python ~/code/Jellyfin-Migrator/jellyfin_migrator/__main__.py
+    """
+    from jellyfin_migrator import core
+    core.main()
